app/requirements.txt
fastapi==0.115.2
feedparser==6.0.11
google-cloud-monitoring>=2.21.0
kubernetes>=29.0.0
firebase-admin>=6.5.0
httpx==0.28.1
jinja2==3.1.4
langchain-core==0.3.77
langchain-google-vertexai==2.1.2
PyYAML==6.0.2
pytest==8.3.3
uvicorn[standard]==0.30.1
